Rough Draft of Enabling Exchange Tooltip Messages
First ever script with basic logic for enabling Exchange tooltips, specifically the External Recipients tooltip. Makes connection to Office365 after specifying credentials and downloads modules, then does basic checking to confirm whether you want to enable it or not.

Still learning the ropes of PS, so this is pretty cruddy (elseifelseifelseifelseifelseifelseifelseifelseifelseif). It works in a pinch!